Support for import-package and export-package on unresolveable types
Developing OSGI bundles with import/export package is currently a PITA.
You cannot use "Organize Imports" or Ctrl-Space to add imports,
because packages need to be exported (from the supplier) and imported
from the consumer first. This slows down development a lot.

This change extends an existing quickfix to be invokeable on
missing/unresolveable types. Additionally, if the type found cannot be
imported, a quickfix is provided to export the type first.